fix: pivoted geo table doesn't have tooltips
